+++ tests.integration/azure-pipelines/windows-robot-integration-test.yml
