+++ tests.unit/test_edk2_ci_build.py
@@ -12,7 +12,7 @@ import os
 import logging
 import shutil
 from importlib import reload
-from edk2toolext.tests.uefi_tree import uefi_tree
+from uefi_tree import uefi_tree
 from edk2toolext.environment import shell_environment
 from edk2toolext.environment import self_describing_environment
 from edk2toolext.environment import version_aggregator
