+++ tests.unit/test_edk2_update.py
@@ -13,7 +13,7 @@ import os
 import logging
 from importlib import reload
 from edk2toolext.environment import shell_environment
-from edk2toolext.tests.uefi_tree import uefi_tree
+from uefi_tree import uefi_tree
 from edk2toolext.environment import self_describing_environment
 from edk2toolext.environment import version_aggregator
