+++ tests.unit/test_git_dependency.py
