+++ tests.unit/test_image_validation.py
