+++ tests.unit/test_repo_resolver.py
