+++ tests.unit/test_self_describing_environment.py
@@ -13,7 +13,7 @@ import git
 import yaml
 import os
 from edk2toolext.environment import self_describing_environment
-from edk2toolext.tests.uefi_tree import uefi_tree
+from uefi_tree import uefi_tree
 from edk2toolext.environment import version_aggregator
