+++ tests.unit/test_shell_environment.py
