+++ tests.unit/test_signtool_signer.py
