+++ tests.unit/test_versioninfo.py
