+++ tests.unit/testdata/secureboot_audit/dbx.bin
