+++ tests.unit/testdata/versioninfo_full_exe.data
