+++ tests.unit/testdata/versioninfo_minimal_exe.data
